requirements.txt
# ─────────────────────────────────────────────
# requirements.txt — Stress Detection System
# ─────────────────────────────────────────────
# Install all dependencies with:
#   pip install -r requirements.txt

# Core ML / Deep Learning
tensorflow>=2.12.0,<2.16.0   # CNN model (CPU-only is fine for inference)
scikit-learn>=1.3.0           # train_test_split, metrics

# Image Processing
opencv-python-headless>=4.8.0 # face detection, preprocessing (headless = no GUI needed)
Pillow>=10.0.0                 # image loading in Streamlit

# Data handling
numpy>=1.24.0
pandas>=2.0.0                  # reading FER-2013 CSV

# Visualisation (used in model.py for training curves)
matplotlib>=3.7.0

# Web UI
streamlit>=1.28.0
